!!! Check extended description for details !!!
Image2image takes the image and returns the image as it would appear on the sphere but is not dynamic. The OpenGL program has the ability to become dynamic. Not sure if this is exactly what we need but just let me know what needs to be changed or revised. The image may need to be cropped for the mapping to work perfectly.